// src/app/trabajadores/interfaces/trabajador.interface.ts
export interface Trabajador {
    id:              number;
    nombre:          string;
    apellido:        string;
    telefono:        number;
    rut:             string;
    fechaNacimiento: string;
    direccion:       Direccion;
    activo:          number;
}

export interface Direccion {
    calle:   string;
    numero:  number;
    comuna?: string;
    nombre?: string;
}